Separate rotkehlchen utils into two files and into own module
