+++ rotkehlchen/binance.py
@@ -17,7 +17,8 @@ from rotkehlchen.logging import RotkehlchenLogsAdapter
 from rotkehlchen.order_formatting import Trade, TradeType
 from rotkehlchen.typing import ApiKey, ApiSecret, FilePath, Timestamp, TradePair
 from rotkehlchen.user_messages import MessagesAggregator
-from rotkehlchen.utils import cache_response_timewise, rlk_jsonloads
+from rotkehlchen.utils.misc import cache_response_timewise
+from rotkehlchen.utils.serialization import rlk_jsonloads
 
 logger = logging.getLogger(__name__)
 log = RotkehlchenLogsAdapter(logger)
