+++ rotkehlchen/csv_exporter.py
@@ -19,7 +19,7 @@ from rotkehlchen.constants.assets import A_ETH
 from rotkehlchen.fval import FVal
 from rotkehlchen.logging import RotkehlchenLogsAdapter, make_sensitive
 from rotkehlchen.typing import EmptyStr, EventType, Fee, FilePath, Timestamp
-from rotkehlchen.utils import taxable_gain_for_sell, tsToDate
+from rotkehlchen.utils.misc import taxable_gain_for_sell, tsToDate
 
 logger = logging.getLogger(__name__)
 log = RotkehlchenLogsAdapter(logger)
