+++ rotkehlchen/poloniex.py
@@ -28,14 +28,8 @@ from rotkehlchen.order_formatting import (
 )
 from rotkehlchen.typing import ApiKey, ApiSecret, BlockchainAsset, FilePath, Timestamp, TradePair
 from rotkehlchen.user_messages import MessagesAggregator
-from rotkehlchen.utils import (
-    cache_response_timewise,
-    createTimeStamp,
-    retry_calls,
-    rlk_jsonloads,
-    rlk_jsonloads_dict,
-    rlk_jsonloads_list,
-)
+from rotkehlchen.utils.misc import cache_response_timewise, createTimeStamp, retry_calls
+from rotkehlchen.utils.serialization import rlk_jsonloads, rlk_jsonloads_dict
 
 logger = logging.getLogger(__name__)
 log = RotkehlchenLogsAdapter(logger)
